+++ R/c.R
@@ -3,12 +3,26 @@
 #' This is useful way of patching together factors from multiple sources
 #' that really should have the same levels but don't.
 #'
-#' @param fs A list of factors
+#' @param fs A factor, or list of factors.
+#' @param ... Individual factors
 #' @export
 #' @examples
-#' fs <- list(factor("a"), factor("b"), factor(c("a", "b")))
+#' fa <- factor("a")
+#' fb <- factor("b")
+#' fab <- factor(c("a", "b"))
+#'
+#' c(fa, fb, fab)
+#' fct_c(fa, fb, fab)
+#'
+#' # You can also pass a list of factors as the first argument
+#' fs <- list(fa, fb, fab)
 #' fct_c(fs)
-fct_c <- function(fs) {
+fct_c <- function(fs, ...) {
+  if (is.list(fs)) {
+    fs <- c(fs, list(...))
+  } else {
+    fs <- c(list(fs), list(...))
+  }
   fs <- check_factor_list(fs)
 
   if (length(fs) == 0) {