+++ NEWS.md
@@ -1,5 +1,7 @@
 # forcats 0.1.1.9000
 
+* `fct_c()` can take either a list of factors or individual factors (#42).
+
 * `lvls_reorder()`, `fct_inorder()`, and `fct_infreq()` gain an `ordered`
    argument, allowing you to override the existing "ordered" status (#54).